fix(charting): handle-new-year
fixes issue when last datapoint is in the first week of the new year (e.g. 2021).  In many cases this datapoint is actually the last week of the previous year (e.g. 2020), so should be included in previous year data (even though the datapoint date is 2021-01-01).  The ChartFactory._get_chart_data method offsets the data correctly, but the current year line is labeled improperly.  Rather than label the chart based on year of last datapoint, I added a function to calculate the current year - adjusting for early year datapoints.